requirements.txt
psycopg2-binary
sqlparse
mariadb
pymysql
